fix(rl): drop the forward autocast wrapper; add --train.force_sync_mode
On routing-sensitive MoE checkpoints, vllm_kl climbs during training because the
tiny per-step actor/rollout weight difference is amplified into top-k routing
flips: torch.autocast forces the fp32-kept MoE gate (gate_precision float32) to
compute in bf16, so its degraded scores disagree with vLLM's fp32 router. Two
independent factors feed the weight difference and either fix alone flattens the
gap: (a) the gate autocast, (b) the async pipeline freeing the rollout slot before
train_step refits vLLM, which generates the next batch one step stale.

- models/base.py: remove the forward-level torch.autocast entirely. FSDP2's
  MixedPrecisionPolicy already casts the bulk params to bf16; the extra wrapper
  only degraded the fp32-kept gate. Drops MOLT_FORWARD_AUTOCAST and
  _forward_autocast_dtype.
- models/critic.py: align the bf16 backbone hidden to the fp32 value head
  explicitly, since there is no forward autocast to do it now.
- cli + rl_trainer: add --train.force_sync_mode, which frees the rollout slot only
  AFTER the refit so the next batch is generated with the weights the trainer
  recomputes it under (strictly on-policy), trading the generate/train overlap.
- fsdp/refit.py: drop the unused should_refit_state_dict_entry /
  state_dict_parameter_trainability (the sender pushes every state_dict entry and
  vLLM's load_weights decides what to accept); remove its obsolete test.
- recipes: drop the now-dead MOLT_FORWARD_AUTOCAST pins.

+++ molt/models/critic.py
@@ -56,14 +56,16 @@ class _ValueHead(nn.Module):
 
     def __init__(self, hidden_size: int):
         super().__init__()
-        # fp32 to match the fp32 master-weight convention; bf16 compute comes from
-        # the same forward autocast the backbone uses.
+        # fp32 to match the fp32 master-weight convention; the value head is tiny,
+        # so running it in fp32 (upcasting the bf16 backbone hidden) is cheap.
         self.proj = nn.Linear(hidden_size, 1, bias=False, dtype=torch.float32)
 
     def forward(self, hidden_states):
         # Gather any TP/SP sharding (no-op for a replicated DTensor or a plain tensor).
         hidden_states = unshard_dtensor(hidden_states)
-        return self.proj(hidden_states)
+        # No forward autocast anymore, so align the bf16 backbone hidden to the
+        # fp32 value-head weight explicitly instead of relying on autocast.
+        return self.proj(hidden_states.to(self.proj.weight.dtype))
 
 
 def _resolve_hidden_size(model) -> int:
